fix: footer refactor and implementation
