+++ packages/chat/src/components/NewFooter/NewFooter.css.ts
@@ -2,10 +2,6 @@ import { keyframes, style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
 import { COLORS } from '@/styles/colors';
-import { PALETTE } from '@/styles/colors.css';
-import { FAMILY } from '@/styles/font';
-import { SIZES } from '@/styles/sizes';
-import { transition } from '@/styles/transitions';
 import { widgetContainer } from '@/views/ChatWidget/styles.css';
 
 const BUTTON_ROW_MARGIN = 10;
@@ -99,62 +95,4 @@ export const messageInputContainer = style({
   backgroundColor: COLORS.white,
   borderTopLeftRadius: '20px',
   borderTopRightRadius: '20px',
-});
-
-export const footerLinksContainer = style({
-  color: COLORS.NEUTRAL_DARK[100],
-  fontFamily: FAMILY,
-  fontSize: '12px',
-  lineHeight: '17px',
-  width: '100%',
-  minHeight: 20,
-  padding: '10px 0',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  gap: '8px',
-});
-
-export const separator = style({
-  height: '2px',
-  width: '2px',
-  borderRadius: SIZES.radius.round,
-  backgroundColor: COLORS.NEUTRAL_LIGHT[500],
-});
-
-export const extraLinkStyles = style({
-  color: COLORS.NEUTRAL_DARK[100],
-  fontFamily: FAMILY,
-  fontSize: '12px',
-  lineHeight: '17px',
-  textDecorationColor: 'transparent',
-  transition: transition(['color', 'text-decoration-color']),
-  selectors: {
-    '&:hover': {
-      textDecoration: 'underline',
-      color: PALETTE.colors[500],
-      textDecorationColor: PALETTE.colors[500],
-    },
-    '&:active': {
-      color: PALETTE.colors[600],
-      textDecorationColor: PALETTE.colors[600],
-    },
-  },
-});
-
-export const voiceflowLink = style({
-  color: COLORS.NEUTRAL_DARK[100],
-  textDecoration: 'underline',
-  textDecorationColor: 'transparent',
-  transition: transition(['color', 'text-decoration-color']),
-  selectors: {
-    '&:hover': {
-      color: PALETTE.colors[500],
-      textDecorationColor: PALETTE.colors[500],
-    },
-    '&:active': {
-      color: PALETTE.colors[600],
-      textDecorationColor: 'transparent',
-    },
-  },
 });
