+++ packages/chat/src/components/VoiceWidget/index.tsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import { Button } from '../Button';
 import { ButtonIcon } from '../Button/ButtonIcon';
 import { ButtonVariant } from '../Button/constants';
+import { BottomLinks } from '../NewFooter/BottomLinks';
 import {
   buttonContent,
   buttonModifier,
@@ -11,11 +12,23 @@ import {
   controlSection,
   titleStyle,
   voiceWidgetContainer,
+  voiceWrapper,
 } from './VoiceWidget.css';
 
-export const VoiceWidget = ({ children }: { children: React.ReactNode }) => {
-  const [isCalling, setIsCalling] = useState(false);
+interface IVoiceWidget {
+  children: React.ReactNode;
+  isListening?: boolean;
+  isTalking?: boolean;
+  footer: {
+    extraLinkText: string;
+    extraLinkUrl: string;
+    showPoweredBy: boolean;
+  };
+}
 
+export const VoiceWidget: React.FC<IVoiceWidget> = ({ children, isListening, isTalking, footer }) => {
+  const [isCalling, setIsCalling] = useState(false);
+  const { showPoweredBy, extraLinkText, extraLinkUrl } = footer;
   const handleButtonClick = () => {
     setIsCalling((prev) => !prev);
     if (isCalling) {
@@ -35,26 +48,38 @@ export const VoiceWidget = ({ children }: { children: React.ReactNode }) => {
     // End call
   };
 
+  let title = '';
+  if (!isCalling) {
+    title = 'How can I help you?';
+  } else if (isListening) {
+    title = 'Listening';
+  } else if (isTalking) {
+    title = 'Talk to interupt';
+  }
+
   return (
-    <div className={voiceWidgetContainer}>
-      <div className={circle}>{children}</div>
-      <div className={controlSection}>
-        <div className={titleStyle}>How can I help you?</div>
-        <Button
-          onClick={handleButtonClick}
-          variant={isCalling ? ButtonVariant.SECONDARY : ButtonVariant.PRIMARY}
-          className={buttonModifier}
-        >
-          <span className={buttonContent({ isVisible: isCalling })}>
-            <ButtonIcon svg="endCall" />
-            <div className={buttonText}>End</div>
-          </span>
-          <span className={buttonContent({ isVisible: !isCalling })}>
-            <ButtonIcon svg="phone" />
-            <div className={buttonText}>Start a call</div>
-          </span>
-        </Button>
+    <div className={voiceWrapper}>
+      <div className={voiceWidgetContainer}>
+        <div className={circle}>{children}</div>
+        <div className={controlSection}>
+          <div className={titleStyle}>{title}</div>
+          <Button
+            onClick={handleButtonClick}
+            variant={isCalling ? ButtonVariant.SECONDARY : ButtonVariant.PRIMARY}
+            className={buttonModifier}
+          >
+            <span className={buttonContent({ isVisible: isCalling })}>
+              <ButtonIcon svg="endCall" />
+              <div className={buttonText}>End</div>
+            </span>
+            <span className={buttonContent({ isVisible: !isCalling })}>
+              <ButtonIcon svg="phone" />
+              <div className={buttonText}>Start a call</div>
+            </span>
+          </Button>
+        </div>
       </div>
+      <BottomLinks extraLinkText={extraLinkText} extraLinkUrl={extraLinkUrl} showPoweredBy={showPoweredBy} />
     </div>
   );
 };
